split-utils * Move functions that are only imported from one file to that file
