+++ src/dashboards/Map/DepartureTag/DepartureTag.tsx
@@ -3,15 +3,22 @@ import { format, isToday } from 'date-fns'
 import { nb } from 'date-fns/locale'
 import { colors } from '@entur/tokens'
 import { Heading4 } from '@entur/typography'
-import { StopPlaceWithDepartures, IconColorType } from '../../../types'
 import {
-    getIcon,
-    getIconColor,
-    getDepartureNumber,
-    getDepartureDirection,
-} from '../../../utils'
+    StopPlaceWithDepartures,
+    IconColorType,
+    LineData,
+} from '../../../types'
+import { getIcon, getIconColor } from '../../../utils'
 import { DepartureIcon } from './DepartureIcon/DepartureIcon'
 import './DepartureTag.scss'
+
+function getDepartureDirection(departure: LineData): string[] {
+    return departure.route.split(/([\s])/g).slice(1)
+}
+
+function getDepartureNumber(departure: LineData): string {
+    return departure.route.split(/[\s]/g)[0] || ''
+}
 
 const DepartureTag = (props: Props): JSX.Element => (
     <div className="departure-tile">
